Created navigation bars for web and mobile view
