+++ src/components/logo/MixLogo.tsx
@@ -6,16 +6,16 @@ const MixLogo = () => {
 			<Image
 				className="dark:hidden"
 				src={"/logo-full-mix-dark.svg"}
-				alt="TheMonkeys Logo"
-				title="TheMonkeys Logo"
+				alt="TheMonkeys"
+				title="TheMonkeys"
 				height={30}
 				width={119}
 			/>
 			<Image
 				className="dark:block hidden"
 				src={"/logo-full-mix-light.svg"}
-				alt="TheMonkeys Logo"
-				title="TheMonkeys Logo"
+				alt="TheMonkeys"
+				title="TheMonkeys"
 				height={30}
 				width={119}
 			/>
